requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
azure-core==1.29.5
azure-ai-formrecognizer==3.3.0
azure-ai-textanalytics==5.3.0
azure-cosmos==4.5.1
azure-search-documents==11.4.0
openai==1.3.0
python-multipart==0.0.6
pydantic==2.4.2
pydantic-settings==2.0.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
httpx==0.25.1 